requirements.txt
Django==4.0.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
asgiref==3.4.1
tzdata==2021.5
sqlparse==0.4.2
pytz==2021.3
django-cors-headers==3.11.0
Pillow==9.0.0